Tests/LibWeb/Text/expected/Editing/execCommand-removeFormat.txt

Div contents: "f<b>o</b>ob<i>a</i>r"
Div contents: "foobar"
